pyproject.toml
[project]
name = "naux-discord-bot"
version = "0.1.0"
description = "Naux - Bot de Discord con IA y funcionalidades de automatización"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.3",
    "python-dotenv>=1.1.1",
    "httpx>=0.28.1",
    "requests>=2.32.5",
    "yt-dlp>=2025.9.26",
    "pynacl>=1.6.0",
]